added confirm delete option

+++ src/app/components/items/items.component.ts
@@ -22,8 +22,10 @@ export class ItemsComponent implements OnInit {
   }
 
   deleteItem(event, item: Item) {
-    this.clearState();
-    this.itemService.deleteItem(item);
+    if (confirm(`Are you sure to delete ${item.title} ? `) == true) {
+      this.clearState();
+      this.itemService.deleteItem(item);
+    }
   }
 
   editItem(event, item: Item) {
